Revert "Merge pull request #17 from slidecart/revert-14-feature/frontend/email-verification-and-reset-password"
This reverts commit 60eec5cdb18fcabf0a992b711da2eb88b7ec845b, reversing
changes made to 493017f1b85b40aa0ebbe68d3b0797febce68c67.

+++ backend/backend/app/src/main/java/com/smartcalender/app/service/OtpService.java
@@ -4,6 +4,7 @@ import com.smartcalender.app.entity.Otp;
 import com.smartcalender.app.repository.OtpRepository;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 
@@ -37,8 +38,11 @@ public class OtpService {
         return otpEntity.getOtp().equals(otp);
     }
 
+    @Transactional
     public void deleteOtp(Long userId) {
-        otpRepository.deleteById(userId);
+        if (otpRepository.existsById(userId)) {
+            otpRepository.deleteById(userId);
+        }
     }
 
     private String generateOtp() {
